refactor: [BE] user 가입/탈퇴 및 권한 조회 관련 기능 전부 auth 모듈 컨트롤러에 추가
