+++ packages/server/src/auth/auth.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
+import { UserModule } from 'src/user/user.module';
 
 @Module({
-  imports: [HttpModule],
+  imports: [HttpModule, UserModule],
   controllers: [AuthController],
   providers: [AuthService],
   exports: [AuthService],
