+++ packages/server/src/user/user.service.ts
@@ -1,12 +1,12 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FindOneOptions, Repository } from 'typeorm';
-import { SignUpDto } from './dto/signup.dto';
+import { SignUpDto } from '../auth/dto/signup.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './entities/user.entity';
 import { Request } from 'express';
 import { AuthService } from 'src/auth/auth.service';
-import { NaverSignInDto } from './dto/naver-singIn.dto';
+import { NaverSignInDto } from '../auth/dto/naver-singIn.dto';
 import { USER_TYPE } from './enum/userType.enum';
 
 @Injectable()
